RABID-525 do full skb copy instead of incrementing refcount
This fix the kernel panic when the original packet leaves the network
stack without a drop. We do not control what happens with the skb
after it leaves our code. The error was that a the dst field was set
to null. Working with a copy should be safer and prevent this types of
errors in the future.