NFLUA-37 add support to iptables LUA target
With this change, programmers can only call the function getpacket
during the target callback. It is not available in the match callback
anymore.
Now, the function getpacket keeps a reference to the original skb
instead of creating a copy. When the function getpacket is called, the
target callback should return NF_STOLEN.