nflua_skb_send: Don't route if destination device is gone
This guards against a null dereference inside route_me_harder (or
ip_route_me_harder, in older kernels).

It isn't clear whether this is the proper fix for this sort of issue, or
even what exactly the issue is, but it seems to occur rarely and having
this fail is better than a panic anyway.