Bump LUAI_HASHLIMIT to 8
As shown in lstring.c:

    /*
    ** Lua will use at most ~(2^LUAI_HASHLIMIT) bytes from a string to
    ** compute its hash
    */
    #if !defined(LUAI_HASHLIMIT)
    #define LUAI_HASHLIMIT     5
    #endif

The default is an inconceivably low 5, which means that only the first
32 characters of string keys in tables are hashed.

For IPv4 that has a maximum length of 40 characters and for IPv6 it can
be up to 88 characters.

Especially for IPv6 this is terrible since srcip alone can be over 32
bytes long. Hence, if we have hundreds of flows for a single srcip,
every lookup for that IP can degrade to a linear search.

Bump it to 8 for safety; we probably won't use keys that are over 256
bytes in length anyway.