nflua: Don't use kallsyms_lookup_name to get ip_forward_options
kallsyms_lookup_name is no longer available, to stop exactly this kind
of bypassing the normal EXPORT_SYMBOL system.

Instead, take a copy of the code we need, which amounts to
ip_forward_options itself plus the ip_rt_get_source helper, which we
modify to only support the output case, as otherwise we'd have to pull
in a lot more code.