nflua: Fix iptables extension init on iptables v1.8.9 and later
libxtables v1.8.9 no longer exports the _init declaration by default
(upstream commit ef108943f69a6e20533d58823740d3f0534ea8ec), we need to
set -DXTABLES_INTERNAL. The name of the macro suggests an assumption
that no iptables extensions are being built externally to the project
itself, which would make us using it like this somewhat dubious, but it
lets us hobble along for now.

-DXTABLES_INTERNAL also brings in an ARRAY_SIZE declaration and exposes
that _init should not be static (it never was in any extension
upstream). Fortunately our ARRAY_SIZE is vestigial and we can just
remove the statics.

Without this change, the issue manifested as e.g. `iptables -m lua`
failing with:

    iptables v1.8.9 (nf_tables): Couldn't load match `lua':No such file or directory

Which was wonderfully confusing.